requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
confluent-kafka==1.3.0
idna==3.4
oauthlib==3.2.2
pip==9.0.1
pkg-resources==0.0.0
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
setuptools==39.0.1
tweepy==4.12.1
urllib3==1.26.13
